relative-ize canvas links outside the course in imports
background information:
----------------------

if you make an API PUT to

 https://example.instructure.com/courses/123/pages/foo

and include links to e.g.

 https://example.instructure.com/courses/456

canvas will rewrite the link in process_incoming_html_content
to look like

 /courses/456

which helps in being able to serve the content from multiple
domains (including, for example, cross-listed courses or
vanity URLs).

when exporting this content, canvas plugs in the primary
account domain to make the link absolute. this is necessary
because the package may be imported into a different domain.

the problem:
-----------

on import, these absolute links are left alone. if the content
is viewed on a secondary domain, the links will appear to be
to external content and navigation will be impeded (the user
may need to log in again, etc.)

test plan:
---------

1. have a canvas + multiple_root_accounts installation, with
   multiple AccountDomains for a single root account. we will
   call these A and B.
2. create two courses under this account, C1 and C2.
3. from domain B, create a page in C1 that links to C2
   (e.g. http://domain-b.local/courses/2)
4. notice in the HTML for the page (as viewed in the Rails console)
   that the link is actually to /courses/2 (at least, it should be;
   this PS hasn't changed this part)
5. export the course
6. import into a new course
7. ensure in the Rails console that the imported content
   also looks like /courses/2, with no domain attached
8. when the content is viewed through the web in domain B,
   the link appears to be in domain B and navigation is unimpeded

flag = none
fixes LS-2390

Change-Id: I767a7722baa06a00a9da46612bc2e1457485b98d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/270541
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Eric Saupe <eric.saupe@instructure.com>
QA-Review: Eric Saupe <eric.saupe@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>