spec: Skip flakey RCE spec
flag=none

This behavior is covered decently in unit tests, but we
should still have an integration-style test to cover it.

MAT-386 is to fix the flakiness of the spec or test the
behavior in a better way.

Test Plan:
specs pass

Change-Id: I1500537b73831166da1bf8a156b763e58764def8
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/271310
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Andrea Cirulli <andrea.cirulli@instructure.com>
QA-Review: Weston Dransfield <wdransfield@instructure.com>
Product-Review: Weston Dransfield <wdransfield@instructure.com>