Allow LTI submissions in new student workflow
Add support for submitting "LTI launch" submissions in the new student
workflow. In the process, the layout for external tools has been
redesigned; although they still only appear for assignments that accept
file uploads, they now appear on the top level alongside the other
submission types. Most of them are behind a "More" button,
aside from a couple select tools like Studio.

Since we can't predict ahead of time whether a given LTI will return a
file or a link, we determine how to handle it when we receive external
content. If a tool returns a file, we add it to the draft's collection
of uploaded files; if a tool returns a link, we show a preview of the
linked content and allow the student to submit it as their submission.

closes EVAL-2000
closes EVAL-1504
flag=assignments_2_student

Test plan:

Setup:
- Have a course with the new student workflow enabled
- Have at least one external tool installed that returns files (like
  Dropbox) and at least one that returns links (like Studio, or the LTI
  test tool)
- Have an assignment that allows file uploads

Actual testing:
- As a student, open the assignment
- Observe the top-level submission type buttons:
  - Studio should have its own button, unless you've renamed it to
    something outlandish
  - There should be a "More" button, which should show additional LTIs
    when clicked
- Launch your file-returning LTI, as noted above
  - When you select a file in the LTI, you should be switched over to
    the file upload panel, and your file should begin uploading
  - Test that this file is saved to the draft when it finishes uploading
    (i.e., it's still there if you reload the page), and you can submit
    the draft
- Start a new attempt
- Launch your link-returning LTI, as noted above
  - After selecting a link in the LTI, you should see your selected
    content rendered in a frame on the page, along with a "Change"
    button that will let you restart the selection process
  - Test that reloading the page preserves this selected content, and
    that you can submit the draft
- As a teacher, open the student's submission page or similar, and check
  that the LTI content appears to you

- Smoke-test the submission workflow more generally to make sure nothing
  broke

Change-Id: I42a8ddd463fa115c4807ded63455cd7d06b4998a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/274808
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>
QA-Review: Adrian Packel <apackel@instructure.com>
Product-Review: Jody Sailor