use fancy midnight relative to the context
Fancy midnight is meant to adjust the date time to the end_of_day if
there is not a time specified. The time of Midnight, or first second of
the day, 00:00:00 is the default when a time is not specified.

Prior behavior if we try to use midnight UTC as a lock_at time, it will
adjust, but was only doing so relative to UTC since it was a time from
the object which came from the db, or from the front end and converted
to UTC since it is stored in UTC.

Now we convert the time to the context time_zone before evaluating in
fancy midnight

test plan
 - set lock_at to midnight UTC
 - it should stay as midnight UTC
 - don't add a time, or set time to midnight local time
 - it should adjust the lock_at to the next day one save

fixes VICE-995
flag=none

Change-Id: I2fe4b53522f5d461ed652cc32e0fa79e8b5a47b4
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/275359
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Caleb Guanzon <cguanzon@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Caleb Guanzon <cguanzon@instructure.com>