Convert QUnit test to jest
I need to edit this file and instead of copying and pasting a test and
modifying one little thing I thought it would be better to clean it up
a bit.

Jest seems to care more about errors than QUnit.
NewKeyModeal#closeModal() explicitly sets `toolConfigurationUrl: null`
which, when rendered, will update it it in NewKeyForm, which was causing
an error because it's required. So I removed the 'required'.

refs INTEROP-7111

Change-Id: I4dd634b32956c8f5602b6c150450968e3545e156
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/278383
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
Reviewed-by: Tucker Mcknight <tmcknight@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>