fix cross-shard launches of course-level LTI tools
The Model.shard(...) form does not properly handle polymorphic
associations *of different types* when not on the same shard as the
current shard. In this case, it doesn't matter that the contexts are on
the same shard as the ContextExternalTool we are looking for; if we are
using ContextExternalTool.shard(...) and we are not currently on that
shard, global IDs will be used instead of local IDs, *if* the types of
the polymorphic association are different (context and course). If

Example where course.global_id == 12340000000001122 and course.shard == 1234
  ContextExternalTool.shard(course.shard).where(context: [course]).active
=>
  SELECT "context_external_tools".* FROM "cluster123_shard_1234"."context_external_tools"
    WHERE "context_external_tools"."context_type" = 'Course'
    AND "context_external_tools"."context_id" = 1122
    AND "context_external_tools"."workflow_state" NOT IN ('deleted', 'disabled')

This also works if we pass in two courses, or two different accounts.
However when we pass in a course and an account, global IDs are
erroneously used.

  ContextExternalTool.shard(course.shard).where(context: [course, course.root_account]).active
=>
  SELECT "context_external_tools".* FROM "cluster123_shard_1234"."context_external_tools"
    WHERE (
       "context_external_tools"."context_type" = 'Course'
       AND "context_external_tools"."context_id" = 12340000000001122
       OR "context_external_tools"."context_type" = 'Account'
       AND "context_external_tools"."context_id" = 12340000000000001
    ) AND "context_external_tools"."workflow_state" NOT IN ('deleted', 'disabled')

This problem is avoided if we just use the shard.activate{} form.

flag=none
fixes INTEROP-6448

Test plan:
This is tricky to test because of the sharded nature. I used a prod
console to test the behavior of find_external_tool() as follows:
- get the course mentioned in the ticket and then use
  ContextExternalTool.find_external_tool() to find the tool with the url
  and course.
- switch to a different shard (e.g. the other shard mentioned in the
  ticket) and repeat. It will return nil.
- monkey-patch find_external_tool by pasting it into the console. repeat
  the last step. It will find the tool.

Change-Id: I08b1acf3cf0c71559dd7f7cd3e7dbed333088453
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/279376
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Mysti Lilla <mysti@instructure.com>
QA-Review: Mysti Lilla <mysti@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>