move blueprint LockManager inside ready callback
because LockManager.init requires the DOM node to be there

test plan: the following pages should render correctly
in a blueprint course and in a blueprint associated course.
the locked banner should appear near the top of the page
in each instance when the content is locked.
 - assignments show
 - assignments edit
 - discussion topics show
 - discussion topics edit
 - quizzes show
 - quizes edit
 - pages show
 - pages edit

refs FOO-2597

Change-Id: I066b4590635efe48fcb690ca14f4ba582bf8f7bc
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/280952
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>