Fix "Copy To" for LTI 1.3 Assignments
Previously, if a user copied an assignment that used an LTI 1.3 Tool,
the tool itself wouldn't get copied over. This meant the assignment
wouldn't launch in the new course. If a teacher then installed the tool
in the new course, then the custom params would be lost, but the tool
would launch. The only workaround was to either copy the assignment
again, or already have the tool installed in the course before copying.
This commit does the latter, ensuring that if an LTI 1.3 assignment is
going to be copied over, the tool is also copied over.

PS: I actually don't think we did support auto tool copying for LTI 1.1
or 2. At least, I didn't see any code in the backend for doing such a
thing. It might be on the frontend when the initial select request is
sent, but I personally think that's the wrong place to put this kind of
logic.

flag=none

fixes INTEROP-7135

test-plan:
* Make sure and set up the LTI 1.3 test tool, but don't install it at an
  account level, just set up the developer key stuff.
* Create two new courses, a source course and a destination course.
* Install the test tool in just the source course, then make a new
  assignment that uses the test tool for its submission type. When
  you're setting up the tool, make sure and add some custom parameters,
  cause we want to see those get copied over.
* Go to the assignment index page, click the three dots, and then copy
  the assignment to the destination course.
* Go to the destination course and ensure that the test tool was
  automatically copied over.
* Launch the copied assignment in the destination course and ensure that
  the custom parameters that you specified earlier also appear in this
  launch.
* Party 🎉!

Change-Id: I3e6021ba79dcf8f627f3629db73d0f71b7871226
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/280963
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Mysti Lilla <mysti@instructure.com>
QA-Review: Mysti Lilla <mysti@instructure.com>
Product-Review: Ryan Hawkins <ryan.hawkins@instructure.com>